Prepare skeleton for api organization
